created jwt service file to separate logic
