--- src/services/apiError.ts
@@ -1,8 +1,0 @@
-export default class ApiError extends Error {
-  statusCode: number;
-
-  constructor(message: string, statusCode: number) {
-    super(message);
-    this.statusCode = statusCode;
-  }
-}
